pyproject.toml
[tool.poetry]
name = "shogi_ai"
version = "0.1.0"
description = ""
authors = ["sansurrogate <takaaki.sanjoh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.9.0"
torchvision = "^0.10.0"
torchaudio = "^0.9.0"
python-shogi = "^1.0.13"
hydra-core = "^1.1.0"
pytorch-lightning = "^1.3.8"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
flake8 = "^3.9.2"
jupyterlab = "^3.0.16"
pandas = "^1.2.5"
black = {version = "^21.6b0", allow-prereleases = true}
matplotlib = "^3.4.2"
tqdm = "^4.61.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
